+++ app/scripts/analytics/analytics.js
@@ -1,0 +1,9 @@
+/*global angular*/
+'use strict';
+
+angular.module('tiwun.analytics', [
+    'tiwun.analytics.factories.SearchItemFactory'
+]);
+
+
+angular.module('tiwun.analytics.factories.SearchItemFactory', []);